test(agent): close peer-job review gaps
